Refactor bill form adding description field
